requirements.txt
streamlit>=1.28.0
openai>=1.0.0
mem0ai>=0.1.0
python-dotenv>=1.0.0
textblob>=0.17.0
openai-whisper>=20231117
